# storage: reduce allocations and iterations in DeleteBucketRange