@zhulongcheng 2019-02-21
The litmus_daily test failed cause  #11965 .