@e-dard 2019-02-21
Hi @zhulongcheng I'm not sure we're going to merge this as we're currently in the process of changing the underlying format for the measurement and field key tags. I will keep you updated, and we _really_ appreciate your continued contributions to the project!